+++ crates/entrypoint/src/cli/command/azure_devops/project/member/list/mod.rs
@@ -1,0 +1,2 @@
+mod azure_devops_project_member_list_cli;
+pub use azure_devops_project_member_list_cli::*;