+++ crates/entrypoint/src/cli/command/azure_devops/project/member/mod.rs
@@ -1,0 +1,3 @@
+mod azure_devops_project_member_cli;
+pub mod list;
+pub use azure_devops_project_member_cli::*;