+++ crates/entrypoint/src/cli/command/azure_devops/project/mod.rs
@@ -1,5 +1,6 @@
 mod azure_devops_project_cli;
 pub mod dump;
 pub mod list;
+pub mod member;
 pub mod show;
 pub use azure_devops_project_cli::*;